requirements.txt
fastapi
pydantic
uvicorn
gunicorn
numpy
torch
torchvision
torchaudio
transformers
load-dotenv
azure-storage-blob
pytz
langchain-openai
langchain
PyPDF2
faiss-cpu
langchain-community